Move _shared files to common/, definitely ARM7-specific stuff to ARM7/
