+++ lib/include/nitro/os/alloc.h
@@ -1,17 +1,10 @@
 #ifndef NITRO_OS_ALLOC_H
 #define NITRO_OS_ALLOC_H
 
-#include <nitro/os/alloc_shared.h>
-#include <nitro/os/arena.h>
+#include <nitro/os/common/alloc_shared.h>
 
-Cell *DLAddFront(Cell* list, Cell* cell);
-Cell *DLExtract(Cell* list, Cell* cell);
-Cell *DLInsert(Cell *original, Cell *inserted);
-void* OS_AllocFromHeap(OSArenaId id, OSHeapHandle heap, u32 size);
-void OS_FreeToHeap(OSArenaId id, OSHeapHandle heap, void* ptr);
-OSHeapHandle OS_SetCurrentHeap(OSArenaId id, OSHeapHandle heap);
-void *OS_InitAlloc(OSArenaId id, void *arenaStart, void *arenaEnd, s32 maxHeaps);
-OSHeapHandle OS_CreateHeap(OSArenaId id, void *start, void *end);
-s32 OS_CheckHeap(OSArenaId id, OSHeapHandle heap);
+#ifdef ARM7
+#include <nitro/os/ARM7/alloc.h>
+#endif
 
 #endif //NITRO_OS_ALLOC_H
