+++ lib/include/nitro/os/common/mutex_shared.h
@@ -13,7 +13,7 @@
 #ifndef NITRO_OS_MUTEX_SHARED_H
 #define NITRO_OS_MUTEX_SHARED_H
 
-#include <nitro/os/thread_shared.h>
+#include <nitro/os/common/thread_shared.h>
 #include <nitro/types.h>
 
 struct OSMutex {
