+++ lib/include/nitro/os/common/reset_shared.h
