+++ lib/include/nitro/os/common/tick_shared.h
@@ -13,7 +13,7 @@
 #ifndef NITRO_OS_TICK_SHARED_H
 #define NITRO_OS_TICK_SHARED_H
 
-#include "nitro/types.h"
+#include <nitro/types.h>
 
 typedef u64 OSTick;
