+++ lib/include/nitro/os/reset.h
@@ -2,7 +2,7 @@
 #define NITRO_OS_RESET_H
 
 #include <nitro/types.h>
-#include <nitro/os/reset_shared.h>
+#include <nitro/os/common/reset_shared.h>
 #include <nitro/pxi/fifo.h>
 
 void OS_InitReset(void);
