+++ lib/include/nitro/os/systemWork.h
@@ -2,6 +2,6 @@
 #define NITRO_OS_SYSTEMWORK_H_
 
 #include <nitro/os/context.h>
-#include <nitro/os/systemWork_shared.h>
+#include <nitro/os/common/systemWork_shared.h>
 
 #endif //NITRO_OS_SYSTEMWORK_H_
