+++ lib/include/nitro/os/tick.h
@@ -1,7 +1,7 @@
 #ifndef NITRO_OS_TICK_H
 #define NITRO_OS_TICK_H
 
-#include <nitro/os/tick_shared.h>
+#include <nitro/os/common/tick_shared.h>
 
 void OS_InitTick(void);
 u16 OS_IsTickAvailable(void);
