+++ lib/include/nitro/os/timer.h
@@ -1,7 +1,7 @@
 #ifndef NITRO_OS_TIMER_H
 #define NITRO_OS_TIMER_H
 
-#include <nitro/os/timer_shared.h>
+#include <nitro/os/common/timer_shared.h>
 #include <nitro/types.h>
 
 void OSi_SetTimerReserved(s32 timerNum);
